Reports TestVector STF test Setup + Codec fixes (#156)
Reports TestVector STF test Setup:
- `VerifyGuarantee_Basic(guarantee)`
- `VerifyGuarantee_RecentHistory(guarantee)`
- `VerifyGuarantee_Authorization(guarantee)`
- `func TestReportVerifyTiny(t *testing.T)` =>`ReportVerify(jsonFile string, exceptErr bool) error `
- `ReportVerify(jsonFile string, exceptErr bool) error ` with JsonParsing and StateTransition Check 

Codec fixes:
1. Due to the original codec encoding method of the map, testing revealed that it still could not encode according to the key's order. Therefore, I declared a KeyValue structure in the map's codec and put it into a Slice for sorting before encoding.
2. Changed the `ExportsCount` in `AvailabilitySpecifier` from `uint32` to `uint16` to match the test vector.
This issue caused the previous work report hash to be incorrect, but it has now been fixed.
3. Changed `map[common.Hash]common.Hash` to use `types.Hash2Hash`, and implemented two types of `MarshalJSON`/`UnmarshalJSON` methods: one using the original dictionary approach, and another using an array of KeyValue pairs.
4. Changed the `jsonkey` of `WorkReport` in `Rho_state` from `workreport` to `report` to match the jsonkey in the test vector.
5. Originally, I optimized the `MarshalJSON`/`UnmarshalJSON` of the `ServiceAccountState` map. However, Shawn discovered that the test vector was updated, and the new test vector has it rewritten as a new structure. Therefore, I removed the original `MarshalJSON`/`UnmarshalJSON` and added the corresponding structure.

TODO:
1. TestReportVerifyTiny two issues
    --- FAIL: TestReportVerifyTiny/dependency_missing-1.json (0.00s)
        report_test.go:240: failed : Expected have error
    --- FAIL: TestReportVerifyTiny/bad_validator_index-1.json (0.00s)
        report_test.go:240: failed : Expected have error

2. rho state transition verification

---------

+++ types/availabilityspecifier.go
@@ -18,7 +18,7 @@ type AvailabilitySpecifier struct {
 	BundleLength        uint32      `json:"length"`
 	ErasureRoot         common.Hash `json:"erasure_root"`
 	ExportedSegmentRoot common.Hash `json:"exports_root"`
-	ExportsCount        uint32      `json:"exports_count"`
+	ExportsCount        uint16      `json:"exports_count"`
 }
 
 // sharing (justified) DA chunks:  Vec<Hash> ++ Blob ++ Vec<Hash> ++ Vec<SegmentChunk> ++ Vec<Hash>.
